Perft of depth 7 is correct

+++ include/movegen.h
@@ -7,20 +7,20 @@
 class MoveGenerator {
 public:
     static void init();
-    static std::vector<u16> generate_legal_moves(ChessGame &);
+    static std::vector<u32> generate_legal_moves(ChessGame &);
 
 private:
-    static void generate_pawn_pushes(std::vector<u16> &, ChessGame &);
-    static void generate_pawn_captures(std::vector<u16> &, ChessGame &);
-    static void generate_knight_moves(std::vector<u16> &, ChessGame &);
-    static void generate_king_moves(std::vector<u16> &, ChessGame &);
-    static void generate_bishop_moves(std::vector<u16> &, ChessGame &);
-    static void generate_rook_moves(std::vector<u16> &, ChessGame &);
-    static void generate_queen_moves(std::vector <u16> &, ChessGame &);
+    static void generate_pawn_pushes(std::vector<u32> &, ChessGame &);
+    static void generate_pawn_captures(std::vector<u32> &, ChessGame &);
+    static void generate_knight_moves(std::vector<u32> &, ChessGame &);
+    static void generate_king_moves(std::vector<u32> &, ChessGame &);
+    static void generate_bishop_moves(std::vector<u32> &, ChessGame &);
+    static void generate_rook_moves(std::vector<u32> &, ChessGame &);
+    static void generate_queen_moves(std::vector <u32> &, ChessGame &);
 
     static void check_detection(u64 *, bool *, bool *, bool);
     static bool simulate_check(u64 *, PieceType, bool, int, int);
-    static std::vector<u16> handle_single_check(std::vector<u16> &, ChessGame &);
+    static std::vector<u32> handle_single_check(std::vector<u32> &, ChessGame &);
 };
 
 #endif // !MOVEGEN_H
